+++ CHANGELOG.md
@@ -4,13 +4,14 @@
 
 ### Features
 
+- Add `firefox` web browser
 - Add `opencode` LLM "agentic" tool
 
 ## v0.4.0 (2026-02-06)
 
 ### Features
 
-- Add `Claude Code` LLM "agentic" tool
+- Add `claude-Code` LLM "agentic" tool
 - Add `codex` LLM "agentic" tool
 
 ## v0.3.0 (2025-12-10)